エラーを解決

+++ Sources/header/BulletManager.h
@@ -13,8 +13,8 @@ public:
 	// ƒvƒŒƒCƒ„[‚²‚Æ‚É‚Â’e‚Ìî•ñ
 	typedef struct BulletPram
 	{
-		float m_BulletStatus[BulletStatus::MAX] = { 1,1,1,1 };
-		int m_BulletType[BulletType::MAX] = { 1,0,0,0,0,0 };
+		float m_BulletStatus[(int)BulletStatus::MAX] = { 1,1,1,1 };
+		int m_BulletType[(int)BulletType::MAX] = { 1,0,0,0,0,0 };
 		std::vector<BaseBullet> m_BulletList;
 	};
